+++ examples/led-blink/app.js
@@ -1,4 +1,4 @@
-const { GPIO, Timer } = require('@makestack/runtime')
+const { GPIO, Timer } = require('makestack')
 const led = new GPIO({ pin: 13, mode: 'out' })
 
 let state = false