+++ sdk/lib/commands/new_plugin.ts
@@ -118,7 +118,7 @@ const TYPESCRIPT_TEMPLATES = [
     {
         filepath: "lib/index.ts",
         template: `\
-const { println } = require('@makestack/runtime')
+const { println } = require('makestack')
 
 export class {{ CamelAppName }} {
     constructor() {