+++ Documentation/changelog.md
@@ -14,7 +14,7 @@ title: Change Log
   - **webui:** Introdue Monaco Editor
   - **sdk:** Remove/replace some commands
 - v0.1.1
-  - apps now have to `require('@makestack/runtime')` to use APIs
+  - apps now have to `require('makestack')` to use APIs
   - rename Store to Config
 - v0.1.0
   - linux: support Raspberry Pi 3 built-in Wi-Fi chip